airflow.cfg configurations
The airflow.cfg file serves as the configuration hub for the Apache Airflow instance, allowing customization of diverse settings and parameters.

I pulled it out of my docker container, and edited the following. 

To declutter the Airflow UI, I disabled the automatic loading of example DAGs, ensuring a cleaner user experience.

Although Celery is the preferred choice for distributed execution in production, I opted for the sequential executor in this project, given its suitability for the task.

To enhance security, I generated a Fernet key, which facilitated the encryption and decryption of sensitive information stored within the Apache Airflow metadata database.

A significant enhancement involved transitioning the metadata database backend from the default SQLite to MySQL, facilitating improved performance and scalability.